+++ src/api.ts
@@ -1,0 +1,31 @@
+export interface ExecutionResult {
+    retc: number;
+    stdout: string | null;
+    stderr: string | null;
+}
+
+
+export interface ExecuteOptions {
+    silent: boolean;
+    environment: Object;
+    collectOutput?: boolean;
+}
+
+export interface CompilationInfo {
+    command: string;
+    directory: string;
+    file: string;
+}
+
+export interface CMakeToolsAPI {
+    sourceDir: string;
+    mainListFile: string;
+    binaryDir: string;
+    cachePath: string;
+    executeCMakeCommand(args: string[], options?: ExecuteOptions): Promise<ExecutionResult>;
+    execute(program: string,
+            args: string[],
+            options?: ExecuteOptions): Promise<ExecutionResult>;
+
+    compilationInfoForFile(filepath: string): Promise<CompilationInfo | null>;
+}