+++ src/cmake.ts
@@ -29,6 +29,8 @@ import {CMakeServerClient,
         CodeModelTarget,
         CodeModelConfiguration
         } from './server-client'
+import {CompilationDatabase} from './compdb';
+import {ExecuteOptions, ExecutionResult, CompilationInfo, CMakeToolsAPI} from './api';
 
 type Maybe<T> = util.Maybe<T>;
 
@@ -131,12 +133,6 @@ export function isTruthy(value: (boolean | string | null | undefined | number)) 
     return !!value;
 }
 
-interface ExecuteOptions {
-    silent: boolean;
-    environment: Object;
-    collectOutput?: boolean;
-};
-
 interface ExecutableTarget {
     name: string;
     path: string;
@@ -152,11 +148,6 @@ export enum EntryType {
     Static,
 };
 
-interface ExecutionResult {
-    retc: number;
-    stdout: Maybe<string>;
-    stderr: Maybe<string>;
-}
 
 interface Test {
     id: number;
@@ -613,7 +604,7 @@ class ThrottledOutputChannel implements vscode.OutputChannel {
     }
 }
 
-export class CMakeTools {
+export class CMakeTools implements CMakeToolsAPI {
     private _context: vscode.ExtensionContext;
     private _channel: vscode.OutputChannel;
     private _ctestChannel: vscode.OutputChannel;
@@ -647,6 +638,7 @@ export class CMakeTools {
     private _workspaceCachePath = path.join(vscode.workspace.rootPath || '~', '.vscode', '.cmaketools.json');
     private _targets: string[] = [];
     private _variantWatcher: vscode.FileSystemWatcher;
+    private _compilationDatabase: Promise<Maybe<CompilationDatabase>> = Promise.resolve(null);
     public os: Maybe<string> = null;
     public systemProcessor: Maybe<string> = null;
     public compilerId: Maybe<string> = null;
@@ -1356,9 +1348,8 @@ export class CMakeTools {
         await this._refreshWorkspaceCacheContent();
 
         // Start loading up available environments early, this may take a few seconds
-        const env_promises = environment.availableEnvironments();
-        for (const pr of env_promises) {
-            pr.then(env => {
+        const env_promises = environment.availableEnvironments().map(
+            pr => pr.then(env => {
                 if (env.variables) {
                     console.log(`Detected available environemt "${env.name}"`);
                     this._availableEnvironments.set(env.name, {
@@ -1369,9 +1360,9 @@ export class CMakeTools {
                 }
             }).catch(e => {
                 debugger;
-                console.log('Error detecting environment', e);
-            });
-        }
+                console.error('Error detecting environment', e);
+            })
+        );
         await Promise.all(env_promises);
         // All environments have been detected, now we can update the UI
         this.activeEnvironments = [];
@@ -1382,7 +1373,7 @@ export class CMakeTools {
             }
         });
 
-        const version_out = (await this.execute(['--version'], {
+        const version_out = (await this.executeCMakeCommand(['--version'], {
             silent: true,
             environment: {},
             collectOutput: true
@@ -1576,6 +1567,14 @@ export class CMakeTools {
             : (entry as LegacyCacheEntry).as<string>();
     }
 
+    public async compilationInfoForFile(filepath: string): Promise<CompilationInfo|null> {
+        const db = await this._compilationDatabase;
+        if (!db) {
+            return null;
+        }
+        return db.getCompilationInfoForUri(vscode.Uri.file(filepath));
+    }
+
     private async _refreshAll() {
         if (!this.serverClient) {
             await this.reloadCMakeCache();
@@ -1583,6 +1582,7 @@ export class CMakeTools {
             await this._reloadMetaData();
         }
         await this._refreshTests();
+        this._compilationDatabase = CompilationDatabase.fromFilePath(path.join(this.binaryDir, 'compile_commands.json'));
     }
 
     private async _refreshCacheContent() {
@@ -1646,7 +1646,7 @@ export class CMakeTools {
         const generator = this.activeGenerator;
         if (generator && /(Unix|MinGW|NMake) Makefiles|Ninja/.test(generator)) {
             const parser = new CMakeTargetListParser();
-            await this.execute(['--build', this.binaryDir, '--target', 'help'], {
+            await this.executeCMakeCommand(['--build', this.binaryDir, '--target', 'help'], {
                 silent: true,
                 environment: {},
             }, parser);
@@ -1656,11 +1656,23 @@ export class CMakeTools {
         return this._targets;
     }
 
+    public replaceVars(str: string): string {
+        const replacements = [
+            ['${buildType}', this.selectedBuildType || 'Unknown'],
+            ['${workspaceRoot}', vscode.workspace.rootPath],
+            ['${workspaceRootFolderName}', path.basename(vscode.workspace.rootPath)]
+        ] as [string, string][];
+        return replacements.reduce(
+            (accdir, [needle, what]) => util.replaceAll(accdir, needle, what),
+            str,
+        );
+    }
+
     /**
      * @brief Read the source directory from the config
      */
     public get sourceDir(): string {
-        const dir = this.config.sourceDirectory.replace('${workspaceRoot}', vscode.workspace.rootPath);
+        const dir = this.replaceVars(this.config.sourceDirectory);
         return util.normalizePath(dir);
     }
 
@@ -1676,9 +1688,7 @@ export class CMakeTools {
      * @brief Get the path to the binary dir
      */
     public get binaryDir(): string {
-        const dir = this.config.buildDirectory
-            .replace('${workspaceRoot}', vscode.workspace.rootPath)
-            .replace('${buildType}', this.selectedBuildType || 'Unknown');
+        const dir = this.replaceVars(this.config.buildDirectory);
         return util.normalizePath(dir, false);
     }
 
@@ -1728,10 +1738,19 @@ export class CMakeTools {
         return (gen && /Visual Studio/.test(gen)) ? 'ALL_BUILD' : 'all';
     }
 
+    public executeCMakeCommand(args: string[],
+                               options: ExecuteOptions = {silent: false, environment: {}},
+                               parser: util.OutputParser = new NullParser)
+    : Promise<ExecutionResult> {
+        console.info('Execute cmake with arguments:', args);
+        return this.execute(this.config.cmakePath, args, options, parser);
+    }
+
     /**
      * @brief Execute a CMake command. Resolves to the result of the execution.
      */
-    public execute(args: string[],
+    public execute(program: string,
+                   args: string[],
                    options: ExecuteOptions = {
                        silent: false,
                        environment: {},
@@ -1741,10 +1760,9 @@ export class CMakeTools {
     : Promise<ExecutionResult> {
         return new Promise<ExecutionResult>(async (resolve) => {
             const silent: boolean = options && options.silent || false;
-            console.info('Execute cmake with arguments:', args);
             let stdout = '';
             let stderr = '';
-            const pipe = proc.spawn(this.config.cmakePath, args, {
+            const pipe = proc.spawn(program, args, {
                 env: Object.assign(
                     {
                         // We set NINJA_STATUS to force Ninja to use the format
@@ -1756,16 +1774,14 @@ export class CMakeTools {
                     this.currentEnvironmentVariables,
                 )
             });
-            const status = msg => vscode.window.setStatusBarMessage(msg, 4000);
             if (!silent) {
                 this.currentChildProcess = pipe;
-                status('Executing CMake...');
                 this._channel.appendLine(
-                    '[vscode] Executing cmake command: cmake '
+                    '[vscode] Executing command: '
                     // We do simple quoting of arguments with spaces.
                     // This is only shown to the user,
                     // and doesn't have to be 100% correct.
-                    + args
+                    + [program].concat(args)
                         .map(a => a.replace('"', '\"'))
                         .map(a => /[ \n\r\f;\t]/.test(a) ? `"${a}"` : a)
                         .join(' ')
@@ -1794,7 +1810,7 @@ export class CMakeTools {
             emitLines(pipe.stderr);
 
             pipe.stdout.on('line', (line: string) => {
-                console.log('cmake [stdout]: ' + line);
+                console.log(program + ' [stdout]: ' + line);
                 if (options.collectOutput) {
                     stdout += line + '\n';
                 }
@@ -1806,10 +1822,10 @@ export class CMakeTools {
                 }
             });
             pipe.stderr.on('line', (line: string) => {
-                console.log('cmake [stderr]: ' + line);
                 if (options.collectOutput) {
                     stderr += line + '\n';
                 }
+                console.log(program + ' [stderr]: ' + line);
                 const progress = parser.parseLine(line);
                 if (!silent) {
                     if (progress)
@@ -1823,7 +1839,7 @@ export class CMakeTools {
                 if (parser instanceof BuildParser) {
                     parser.fillDiagnosticCollection(this._diagnostics);
                 }
-                console.log('cmake exited with return code ' + retc);
+                console.log(program + ' exited with return code ' + retc);
                 if (silent) {
                     resolve({
                         retc: retc,
@@ -1832,12 +1848,13 @@ export class CMakeTools {
                     });
                     return;
                 }
-                this._channel.appendLine('[vscode] CMake exited with status ' + retc);
+                const msg = `${program} existed with status ${retc}`;
+                this._channel.appendLine('[vscode] ' + msg);
                 if (retc !== null) {
-                    status('CMake exited with status ' + retc);
+                    vscode.window.setStatusBarMessage(msg, 4000);
                     if (retc !== 0) {
                         this._warningMessage.color = 'yellow';
-                        this._warningMessage.text = `$(alert) CMake failed with status ${retc}. See CMake/Build output for details`;
+                        this._warningMessage.text = `$(alert) ${program} failed with status ${retc}. See CMake/Build output for details`;
                         this._warningMessage.show();
                         setTimeout(() => this._warningMessage.hide(), 5000);
                     }
@@ -2059,10 +2076,8 @@ export class CMakeTools {
             }
             if (typeof(value) === 'string') {
                 typestr = 'STRING';
-                value = (value as string)
-                    .replace(';', '\\;')
-                    .replace('${workspaceRoot}', vscode.workspace.rootPath)
-                    .replace('${buildType}', this.selectedBuildType || 'Unknown');
+                value = this.replaceVars(value)
+                value = util.replaceAll(value, ';', '\\;');
             }
             if (value instanceof Number || typeof value === 'number') {
                 typestr = 'STRING';
@@ -2078,9 +2093,7 @@ export class CMakeTools {
         await util.writeFile(init_cache_path, initial_cache_content.join('\n'));
         let prefix = this.config.installPrefix;
         if (prefix && prefix !== "") {
-            prefix = prefix
-                .replace('${workspaceRoot}', vscode.workspace.rootPath)
-                .replace('${buildType}', this.selectedBuildType || 'Unknown');
+            prefix = this.replaceVars(prefix);
             settings_args.push("-DCMAKE_INSTALL_PREFIX=" + prefix);
         }
 
@@ -2111,7 +2124,7 @@ export class CMakeTools {
             vscode.window.showErrorMessage('A CMake task is already running. Stop it before trying to configure.');
             return -1;
         }
-        const result = await this.execute(
+        const result = await this.executeCMakeCommand(
             ['-H' + this.sourceDir.replace(/\\/g, path.posix.sep),
              '-B' + this.binaryDir.replace(/\\/g, path.posix.sep),
              '-C' + init_cache_path]
@@ -2188,7 +2201,7 @@ export class CMakeTools {
         })();
         this._channel.show();
         this.statusMessage = 'Building...';
-        const result = await this.execute([
+        const result = await this.executeCMakeCommand([
             '--build', this.binaryDir,
             '--target', target,
             '--config', this.selectedBuildType || 'Debug',
@@ -2392,7 +2405,7 @@ export class CMakeTools {
             return build_retc;
         }
         const retc = (
-            await this.execute(
+            await this.executeCMakeCommand(
                 [
                     '-E', 'chdir', this.binaryDir,
                     'ctest', '-j' + this.numCTestJobs,